+++ jdbc/src/main/java/org/springframework/transaction/core/TransactionTemplate.java
@@ -1,5 +1,9 @@
 package org.springframework.transaction.core;
 
+import org.springframework.transaction.support.ConnectionHolder;
+import org.springframework.transaction.support.TransactionExecutor;
+import org.springframework.transaction.support.TransactionManager;
+
 import javax.sql.DataSource;
 import java.sql.Connection;
 
@@ -8,7 +12,7 @@ public class TransactionTemplate {
     private final TransactionManager transactionManager;
 
     public TransactionTemplate(DataSource dataSource) {
-        this.transactionManager = new TransactionManager(dataSource);
+        this.transactionManager = new TransactionManager(dataSource, ConnectionHolder.getInstance());
     }
 
     public <T> T execute(TransactionExecutor<T> transactionExecutor) {
