+++ jdbc/src/main/java/org/springframework/transaction/support/TransactionExecutor.java
@@ -1,4 +1,4 @@
-package org.springframework.transaction.core;
+package org.springframework.transaction.support;
 
 import java.sql.Connection;
